requirements.txt
requests
beautifulsoup4
groq
schedule
lxml
tqdm
python-dotenv
